move utf BOM stripping logic to StringUtils module and test in model and form objects
