+++ app/models/digital_object/asset.rb
@@ -6,6 +6,7 @@ class DigitalObject::Asset < DigitalObject::Base
   include DigitalObject::Assets::Transcript
   include DigitalObject::Assets::IndexDocument
   include DigitalObject::Assets::Captions
+  include Hyacinth::Utils::StringUtils
 
   UNKNOWN_DC_TYPE = 'Unknown'
   VALID_DC_TYPES = [UNKNOWN_DC_TYPE, 'Dataset', 'MovingImage', 'Software', 'Sound', 'StillImage', 'Text']
@@ -385,51 +386,5 @@ class DigitalObject::Asset < DigitalObject::Base
         'errors' => errors.full_messages
       }
     json
-  end
-
-  def self.ints_as_binstring(ints = [])
-    stub = "".force_encoding(Encoding::ASCII_8BIT)
-    ints.inject(stub) { |m,c| m << c }
-  end
-
-  BOM_UTF_8 = ints_as_binstring([239,187,191])
-  BOM_UTF_16BE = ints_as_binstring([254,255])
-  BOM_UTF_16LE = ints_as_binstring([255,254])
-  BOM_UTF_32BE = ints_as_binstring([0,0,254,255])
-  BOM_UTF_32LE = ints_as_binstring([255,254,0,0])
-
-
-  def encoded_string(source, target_encoding = Encoding::UTF_8)
-    magic_bytes = source.each_byte.lazy.first(4)
-    magic_bytes = DigitalObject::Asset.ints_as_binstring(magic_bytes)
-    return source if source.encoding == target_encoding
-    # these codepoint patterns indicate unicode data in an ASCII_8 string
-    case
-    when magic_bytes.start_with?(BOM_UTF_16BE)
-      source_encoding = Encoding::UTF_16BE
-      source = source.byteslice(BOM_UTF_16BE.length..-1)
-    when magic_bytes.start_with?(BOM_UTF_16LE)
-      source_encoding = Encoding::UTF_16LE
-      source = source.byteslice(BOM_UTF_16LE.length..-1)
-    when magic_bytes.start_with?(BOM_UTF_8)
-      source_encoding = Encoding::UTF_8
-      source = source.byteslice(BOM_UTF_8.length..-1)
-    when magic_bytes.start_with?(BOM_UTF_32BE)
-      source_encoding = Encoding::UTF_32BE
-      source = source.byteslice(BOM_UTF_32BE.length..-1)
-    when magic_bytes.start_with?(BOM_UTF_32LE)
-      source_encoding = Encoding::UTF_32LE
-      source = source.byteslice(BOM_UTF_32LE.length..-1)
-    else
-      detection = CharlockHolmes::EncodingDetector.detect(source)
-      source_encoding = detection[:encoding] ?
-        Encoding.find(detection[:encoding]) : Encoding::ASCII_8BIT
-    end
-    if source.encoding == target_encoding
-      # just the trim, thanks
-      source
-    else
-      source.force_encoding(source_encoding).encode(target_encoding)
-    end
   end
 end
